Extract code mirror specific code to a separate addon
